chore: remove radio db table and improve performance page
