+++ requirements.txt
@@ -24,6 +24,8 @@ pyright==1.1.393
     # via -r requirements.in
 pytest==8.3.4
     # via -r requirements.in
+pyyaml==6.0.2
+    # via -r requirements.in
 requests==2.32.3
     # via -r requirements.in
 ruff==0.9.5
